collections

+++ invoicing/src/main/java/com/kynsoft/finamer/invoicing/domain/dto/CollectionSummaryDto.java
@@ -6,18 +6,18 @@ import lombok.Getter;
 @Builder
 @Getter
 public class CollectionSummaryDto {
-    private int totalInvoiceWith30;
+    private long totalInvoiceWith30;
     private double totalInvoiceBalanceWithAging30;
-    private int totalInvoicePercentWithAging30;
-    private int totalInvoiceWithAging60;
+    private long totalInvoicePercentWithAging30;
+    private long totalInvoiceWithAging60;
     private double totalInvoiceBalanceWithAging60;
-    private int totalInvoicePercentWithAging60;
-    private int totalInvoiceWithAging90;
+    private long totalInvoicePercentWithAging60;
+    private long totalInvoiceWithAging90;
     private double totalInvoiceBalanceWithAging90;
-    private int totalInvoicePercentWithAging90;
-    private int totalInvoiceWithAging120;
+    private long totalInvoicePercentWithAging90;
+    private long totalInvoiceWithAging120;
     private double totalInvoiceBalanceWithAging120;
-    private int totalInvoicePercentWithAging120;
-    private int totalInvoiceWithAging0;
+    private long totalInvoicePercentWithAging120;
+    private long totalInvoiceWithAging0;
     private double totalInvoiceAmount;
 }
